Finished game - maybe update later probably not

+++ Coordinate.cs
@@ -1,6 +1,4 @@
-using System.Diagnostics.CodeAnalysis;
-
 public struct Coordinate{
-    public int X {get; set;}
-    public int Y {get; set;}
+    public byte X {get; set;}
+    public byte Y {get; set;}
 }
